requirements.txt
colorama==0.4.6
loguru==0.7.3
pyaes==1.6.1
pyasn1==0.6.2
python-dotenv==1.2.1
rsa==4.9.1
Telethon==1.42.0
win32_setctime==1.2.0
